// src/app/model/episodes.ts
export interface Episode {
  id: number
  name: string
  air_date: string
  episode: string
  characters: string[]
  url: string
  created: string
  info: Info
  results: Result[]
}

export interface Info {
  count: number
  pages: number
  next: string
  prev: any
}

export interface Result {
  id: number
  name: string
  air_date: string
  episode: string
  characters: string[]
  url: string
  created: string
}